feat: Modify the HTTP API amount parameter to the smallest unit of bitcoin. Prevent loss of precision.

+++ utils/btc/struct.go
@@ -45,12 +45,12 @@ type Input struct {
 	WIF          string
 	RedeemScript string
 	SegWit       bool
-	Amount       float64
+	Amount       int64
 }
 
 type Output struct {
 	PayToAddress string
-	Amount       float64
+	Amount       int64
 }
 
 type Inputs []Input
